+++ client/modules/IDE/components/AddToCollectionList.jsx
@@ -10,7 +10,7 @@ import * as ProjectsActions from '../actions/projects';
 import * as CollectionsActions from '../actions/collections';
 import * as ToastActions from '../actions/toast';
 import getSortedCollections from '../selectors/collections';
-import Item from './QuickAddList';
+import Item from './AddToListRow';
 import Table from './Table';
 
 const DIRECTION = {
