+++ client/modules/IDE/components/AddToListRow/Icons.jsx
