+++ client/modules/IDE/components/AddToListRow/index.js
@@ -1,0 +1,1 @@
+export { default } from './AddToListRow.jsx';
