--- client/modules/IDE/components/QuickAddList/index.js
@@ -1,1 +1,0 @@
-export { default } from './QuickAddList.jsx';
